+++ include/questao1/fruta.h
@@ -3,7 +3,7 @@
  * @brief	Definicao da classe Fruta para representar um fruta
  * @author	Lael Rodrigues(laelrodrigues7@gmail.com)
  * @since	25/10/2017
- * @data	26/10/2017
+ * @date	29/10/2017
  */
 
 #ifndef _FRUTA_H_
@@ -14,9 +14,9 @@
 
 /**
  * @class 	Fruta fruta.h
- * @brief	Classe que representa um produto
+ * @brief	Classe que representa um fruta
  * @details Os atributos do produto: codigo, descricao,
- *  		preco, data e validade			
+ *  		preco, data do lote e validade			
  */
 
 class Fruta : public Produto
@@ -27,28 +27,28 @@ public:
 	Fruta();
 
 	/** @brief Construtor parametrizado */
-	Fruta(std::string _codigo, std::string _descricao, short _preco, 
-			std::string _data, short _validade);
+	Fruta(std::string _codigo, std::string _descricao, double _preco, 
+			std::string _data, int _validade);
 
 	/** @brief Destrutor padrao */
 	~Fruta();
 
 private:
 	std::string m_data_lote; 	/**< Data do lote */
-	short m_validade;			/**< Validade */	
+	int m_validade;			/**< Validade */	
 
 public:
 	/** @brief Retorna a data do lote */
 	std::string getDataLote();
 
 	/** @brief Retorna a validade */
-	short getValidade();
+	int getValidade();
 
 	/** @brief Modifica a data do lote */
 	void setDataLote(std::string _data);
 
 	/** @brief Modifica a validade */
-	void setValidade(short _validade);
+	void setValidade(int _validade);
 
 private:
 	/** @brief Metodo para imprimir os dados de uma fruta */
