+++ src/questao1/bebida.cpp
@@ -4,7 +4,7 @@
  *			para representar uma bebida
  * @author	Lael Rodrigues(laelrodrigues7@gmail.com)
  * @since	25/10/2017
- * @date	25/10/2017
+ * @date	26/10/2017
  */
 
 #include <iomanip>
@@ -19,7 +19,7 @@ Bebida::Bebida() {}
  * @param _preco Preco da bebida
  * @param _teor_alcoolico Teor alcoolico da bebida  
  */
-Bebida::Bebida(std::string _codigo, std::string _descricao, short _preco, 
+Bebida::Bebida(std::string _codigo, std::string _descricao, double _preco, 
 	int _teor_alcoolico):
 	Produto(_codigo, _descricao, _preco), teor_alcoolico(_teor_alcoolico) {}
